+++ core/src/main/java/nl/nn/adapterframework/validation/XercesXmlValidator.java
@@ -329,6 +329,7 @@ public class XercesXmlValidator extends AbstractXmlValidator {
 		// (Wsdl)XmlValidator (using
 		// mergeXsdsGroupedByNamespaceToSchemasWithoutIncludes)).
 		// See comment in method XmlValidator.getSchemas() too.
+		// See ClassLoaderXmlEntityResolver too.
 		return new XMLInputSource(null, schema.getSystemId(), null, schema.getInputStream(), null);
 	}